requirements.txt
Flask==0.12.2
gunicorn==19.7.1
gcloud
firebase_admin
google-cloud-firestore